+++ lib/src/endpoints.rs
@@ -82,6 +82,7 @@ pub fn default_endpoints() -> Vec<Endpoint> {
         plugins::search::search_endpoint(),
         plugins::files::upload_endpoint(),
         plugins::files::download_endpoint(),
+        plugins::export::export_endpoint(),
         #[cfg(feature = "html")]
         plugins::bookmark::bookmark_endpoint(),
         plugins::importer::import_endpoint(),