+++ server/src/routes.rs
@@ -20,6 +20,7 @@ include!(concat!(env!("OUT_DIR"), "/generated.rs"));
 pub fn config_routes(app: &mut actix_web::web::ServiceConfig) {
     app.service(web::resource("/ws").to(handlers::web_sockets::web_socket_handler))
         .service(web::resource("/download/{path:[^{}]+}").to(handlers::download::handle_download))
+        .service(web::resource("/export").to(handlers::export::handle_export))
         // This `generate` imports the static files from the `app_assets` folder
         .service(
             ResourceFiles::new("/", generate())